# backend/swagger.yaml
openapi: 3.0.3
info:
  title: Lead Engine CRE API
  description: |
    REST API for the Lead Engine CRE marketplace — a decentralized lead exchange with
    sealed-bid auctions, ZK-verified compliance, and x402 USDC settlement.

    ## Authentication
    All endpoints require a JWT Bearer token obtained via the wallet auth flow:
    1. `GET /api/v1/auth/nonce/{address}` — get signing message
    2. `POST /api/v1/auth/wallet` — submit signature, receive JWT
    3. Include `Authorization: Bearer <token>` on all subsequent requests

    ## Lead Sources
    Leads can be submitted via three channels:
    - **PLATFORM** — the built-in web form
    - **API** — programmatic submission (this endpoint)
    - **OFFSITE** — hosted landers and webhook integrations
  version: 1.0.0
  contact:
    name: Lead Engine CRE
    url: https://github.com/bnmbnmai/lead-engine-cre
servers:
  - url: http://localhost:3001
    description: Local development
  - url: https://lead-engine-cre-api.onrender.com
    description: Production
...
paths:
...
  /api/v1/leads/submit:
    post:
      tags: [Leads]
      summary: Submit a lead
      description: |
        Submit a new lead to the marketplace. The lead is automatically verified by
        the CRE (Compliance & Risk Engine), matched against active asks, and enters
        auction if a match is found.

        **Hybrid sources**: Use `source: "API"` for programmatic submissions or
        `source: "OFFSITE"` for webhook/lander integrations.

        **Geo-mismatch handling**: If no active ask matches the lead's geo, the lead
        is created but won't enter auction until a matching ask appears. It expires
        after `expiresInMinutes`.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LeadSubmit"
            examples:
              roofing_lead:
                summary: Roofing lead from FL (US)
                value:
                  vertical: roofing
                  source: API
                  geo:
                    country: US
                    state: FL
                    zip: "33101"
                  reservePrice: 35.00
                  expiresInMinutes: 60
                  tcpaConsentAt: "2026-02-09T05:00:00Z"
                  parameters:
                    roof_type: shingle
                    damage_type: storm
                    insurance_claim: true
              solar_lead_brazil:
                summary: Solar lead from São Paulo (BR)
                value:
                  vertical: solar
                  source: API
                  geo:
                    country: BR
                    state: SP
                    city: São Paulo
                    zip: "01310-100"
                  reservePrice: 25.00
                  expiresInMinutes: 120
                  tcpaConsentAt: "2026-02-09T05:00:00Z"
                  parameters:
                    monthly_bill: 350
                    ownership: own
                    panel_interest: purchase
              b2b_lead_kenya:
                summary: B2B SaaS lead from Nairobi (KE)
                value:
                  vertical: b2b_saas
                  source: PLATFORM
                  geo:
                    country: KE
                    state: NBO
                    city: Nairobi
                    zip: "00100"
                  reservePrice: 40.00
                  expiresInMinutes: 60
                  parameters:
                    company_size: "11-50"
                    industry: tech
                    budget_range: "5k-25k"
              mortgage_lead:
                summary: Mortgage lead from CA (US)
                value:
                  vertical: mortgage
                  source: API
                  geo:
                    country: US
                    state: CA
                    city: Los Angeles
                    zip: "90001"
                  reservePrice: 50.00
                  expiresInMinutes: 120
                  tcpaConsentAt: "2026-02-09T05:00:00Z"
                  parameters:
                    loan_type: purchase
                    credit_range: good_700-749
                    property_type: single_family
                    purchase_price: 650000
      responses:
        "201":
          description: Lead created and optionally matched
          content:
            application/json:
              schema:
                type: object
                properties:
                  lead:
                    type: object
                    properties:
                      id:
                        type: string
                      vertical:
                        type: string
                      status:
                        type: string
                        enum: [PENDING_AUCTION, IN_AUCTION]
                      isVerified:
                        type: boolean
                      matchingAsks:
                        type: integer
                      auctionEndAt:
                        type: string
                        format: date-time
                        nullable: true
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "422":
          description: Lead failed CRE verification
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                  details:
                    type: object
  /api/v1/leads:
    get:
      tags: [Leads]
      summary: List leads
      security:
        - bearerAuth: []
      parameters:
        - name: vertical
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
            enum: [PENDING_AUCTION, IN_AUCTION, REVEAL_PHASE, SOLD, EXPIRED]
        - name: state
          in: query
          schema:
            type: string
            minLength: 2
            maxLength: 2
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            minimum: 0
            default: 0
        - name: sortBy
          in: query
          schema:
            type: string
            enum: [createdAt, reservePrice, auctionEndAt]
            default: createdAt
        - name: sortOrder
          in: query
          schema:
            type: string
            enum: [asc, desc]
            default: desc
      responses:
        "200":
          description: Paginated list of leads
          content:
            application/json:
              schema:
                type: object
                properties:
                  leads:
                    type: array
                    items:
                      type: object
                  pagination:
                    $ref: "#/components/schemas/Pagination"
  /api/v1/leads/{id}:
    get:
      tags: [Leads]
      summary: Get lead details
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Lead details with bids and verification
        "404":
          description: Lead not found

  # ─── Asks ───────────────────────────────
  /api/v1/asks:
    get:
      tags: [Asks]
      summary: List asks
      security:
        - bearerAuth: []
      parameters:
        - name: vertical
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
            enum: [ACTIVE, PAUSED, EXPIRED, CANCELLED]
        - name: minPrice
          in: query
          schema:
            type: number
        - name: maxPrice
          in: query
          schema:
            type: number
        - name: state
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: Paginated ask listings
          content:
            application/json:
              schema:
                type: object
                properties:
                  asks:
                    type: array
                    items:
                      $ref: "#/components/schemas/Ask"
                  pagination:
                    $ref: "#/components/schemas/Pagination"

    post:
      tags: [Asks]
      summary: Create ask
      description: Create a new ask listing. Requires seller role.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AskCreate"
      responses:
        "201":
          description: Ask created
          content:
            application/json:
              schema:
                type: object
                properties:
                  ask:
                    $ref: "#/components/schemas/Ask"
  /api/v1/asks/{id}:
    get:
      tags: [Asks]
      summary: Get ask details
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Ask with seller info and leads
        "404":
          description: Ask not found

  # ─── Bids ───────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    LeadSubmit:
      type: object
      required: [vertical, geo]
      properties:
        vertical:
          type: string
          enum: [solar, mortgage, roofing, insurance, home_services, b2b_saas, real_estate, auto, legal, financial]
        geo:
          $ref: "#/components/schemas/Geo"
        source:
          type: string
          enum: [PLATFORM, API, OFFSITE]
          default: PLATFORM
        reservePrice:
          type: number
          minimum: 0.01
        tcpaConsentAt:
          type: string
          format: date-time
        consentProof:
          type: string
        encryptedData:
          type: string
        dataHash:
          type: string
        expiresInMinutes:
          type: integer
          minimum: 5
          maximum: 10080
          default: 60
        parameters:
          type: object
          description: |
            Vertical-specific fields. Examples:
            - **roofing**: roof_type, damage_type, insurance_claim, roof_age, square_footage
            - **mortgage**: loan_type, credit_range, property_type, purchase_price, down_payment_pct
            - **solar**: roof_age, monthly_bill, ownership, panel_interest, shade_level
            - **insurance**: coverage_type, current_provider, policy_expiry, num_drivers
          additionalProperties: true
    Geo:
      type: object
      properties:
        country:
          type: string
          description: ISO 3166-1 alpha-2 country code
          minLength: 2
          maxLength: 2
          default: US
          example: US
        state:
          type: string
          description: US state, CA province, AU state, etc.
          maxLength: 4
        region:
          type: string
          description: Free-text region (for countries without predefined state lists)
          maxLength: 100
        city:
          type: string
        zip:
          type: string
          description: Postal code (US ZIP, UK postcode, CA postal code, etc.)
          pattern: "^[A-Z0-9 -]{2,10}$"
        geoHash:
          type: string
        lat:
          type: number
          minimum: -90
          maximum: 90
        lng:
          type: number
          minimum: -180
          maximum: 180
    AskCreate:
      type: object
      required: [vertical, geoTargets, reservePrice]
      properties:
        vertical:
          type: string
        geoTargets:
          type: object
          properties:
            country:
              type: string
              description: ISO 3166-1 alpha-2 country code (default US)
              minLength: 2
              maxLength: 2
              default: US
            regions:
              type: array
              description: Region/state/province codes for the selected country
              items:
                type: string
            states:
              type: array
              deprecated: true
              description: "Deprecated: use `regions` instead. Alias for backward compatibility."
              items:
                type: string
            zips:
              type: array
              items:
                type: string
            excludeZips:
              type: array
              items:
                type: string
            radius:
              type: object
              properties:
                lat:
                  type: number
                lng:
                  type: number
                miles:
                  type: number
        reservePrice:
          type: number
          minimum: 0.01
        buyNowPrice:
          type: number
        acceptOffSite:
          type: boolean
          default: true
        auctionDuration:
          type: integer
          minimum: 300
          maximum: 604800
          default: 3600
        revealWindow:
          type: integer
          minimum: 60
          maximum: 3600
          default: 900
        expiresInDays:
          type: integer
          minimum: 1
          maximum: 90
          default: 30
        parameters:
          type: object
          additionalProperties: true
    Ask:
      type: object
      properties:
        id:
          type: string
        vertical:
          type: string
        geoTargets:
          type: object
        reservePrice:
          type: number
        buyNowPrice:
          type: number
          nullable: true
        status:
          type: string
          enum: [ACTIVE, PAUSED, EXPIRED, CANCELLED]
        acceptOffSite:
          type: boolean
        expiresAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
...
    Pagination:
      type: object
      properties:
        total:
          type: integer
        limit:
          type: integer
        offset:
          type: integer
        hasMore:
          type: boolean
    Error:
      type: object
      properties:
        error:
          type: string
        details:
          type: array
          items:
            type: object
...